# openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "description": "One envelope for every error response: `{ \"code\", \"message\" }`.",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "CreateDraftRequest": {
        "properties": {
          "allow_fallback": {
            "type": "boolean"
          },
          "source": {
            "$ref": "#/components/schemas/DraftSourceDto"
          },
          "tier": {
            "$ref": "#/components/schemas/DraftTierDto"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "topics",
          "tier",
          "source"
        ],
        "type": "object"
      },
...
      "DraftDto": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "fallback_from": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DraftSourceDto"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "publish_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "publish_stage": {
            "type": [
              "string",
              "null"
            ]
          },
          "published_market_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "$ref": "#/components/schemas/DraftSourceDto"
          },
          "spec": {
            "$ref": "#/components/schemas/DraftSpecDto"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "spec",
          "source",
          "status",
          "expires_at",
          "created_at"
        ],
        "type": "object"
      },
      "DraftSourceDto": {
        "enum": [
          "template",
          "llm"
        ],
        "type": "string"
      },
      "DraftSpecDto": {
        "properties": {
          "description": {
            "type": "string"
          },
          "fee_bps": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "hidden_window_secs": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "min_votes_to_resolve": {
            "format": "int32",
            "type": "integer"
          },
          "open_secs": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "question": {
            "type": "string"
          },
          "seed_micro": {
            "format": "int64",
            "type": "integer"
          },
          "slug": {
            "type": "string"
          },
          "tier": {
            "$ref": "#/components/schemas/DraftTierDto"
          },
          "video_script": {
            "type": "string"
          }
        },
        "required": [
          "question",
          "description",
          "video_script",
          "slug",
          "tier",
          "seed_micro",
          "fee_bps",
          "min_votes_to_resolve",
          "open_secs",
          "hidden_window_secs"
        ],
        "type": "object"
      },
      "DraftTierDto": {
        "enum": [
          "daily",
          "flash"
        ],
        "type": "string"
      },
...
      "EditDraftRequest": {
        "properties": {
          "reviewer_id": {
            "format": "uuid",
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/DraftSpecDto"
          }
        },
        "required": [
          "reviewer_id",
          "spec"
        ],
        "type": "object"
      },
...
      "PublishStatusDto": {
        "description": "`GET /admin/drafts/{id}/publish_status` — the 202 status URL target for\nthe command-based publish_now (codex r3 NEW-4).",
        "properties": {
          "draft_id": {
            "format": "uuid",
            "type": "string"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "result_market_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "status_url": {
            "description": "Poll here (also returned by the 202 publish-now authorization).",
            "type": "string"
          }
        },
        "required": [
          "draft_id",
          "status",
          "status_url"
        ],
        "type": "object"
      },
...
      "ReviewDraftRequest": {
        "properties": {
          "reviewer_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "reviewer_id"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "adapters",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/drafts": {
      "get": {
        "operationId": "list_drafts",
        "parameters": [
          {
            "description": "maximum drafts, 1..=500",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DraftDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": "invalid admin token"
          },
          "422": {
            "description": "invalid limit"
          }
        },
        "tags": [
          "content"
        ]
      },
      "post": {
        "operationId": "create_drafts",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DraftDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": "invalid admin token"
          },
          "409": {
            "description": "draft admission limit reached"
          },
          "422": {
            "description": "invalid draft request"
          }
        },
        "tags": [
          "content"
        ]
      }
    },
    "/admin/drafts/{id}": {
      "patch": {
        "operationId": "edit_draft",
        "parameters": [
          {
            "description": "draft id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": "invalid admin token"
          },
          "404": {
            "description": "unknown draft"
          },
          "409": {
            "description": "draft is not editable"
          }
        },
        "tags": [
          "content"
        ]
      }
    },
    "/admin/drafts/{id}/approve": {
      "post": {
        "operationId": "approve_draft",
        "parameters": [
          {
            "description": "draft id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": "invalid admin token"
          },
          "404": {
            "description": "unknown draft"
          },
          "409": {
            "description": "draft cannot be approved"
          }
        },
        "tags": [
          "content"
        ]
      }
    },
    "/admin/drafts/{id}/publish_now": {
      "post": {
        "operationId": "publish_now",
        "parameters": [
          {
            "description": "draft id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishStatusDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": "invalid admin token"
          },
          "404": {
            "description": "unknown draft"
          },
          "409": {
            "description": "draft is not approved"
          }
        },
        "tags": [
          "content"
        ]
      }
    },
    "/admin/drafts/{id}/publish_status": {
      "get": {
        "operationId": "publish_status",
        "parameters": [
          {
            "description": "Draft id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishStatusDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "ops"
        ]
      }
    },
    "/admin/drafts/{id}/reject": {
      "post": {
        "operationId": "reject_draft",
        "parameters": [
          {
            "description": "draft id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": "invalid admin token"
          },
          "404": {
            "description": "unknown draft"
          },
          "409": {
            "description": "draft cannot be rejected"
          }
        },
        "tags": [
          "content"
        ]
      }
    },
...
  },
...
}